Renamed ItemFeed to FeedComponent
